update: init_global_conf
