+++ yatori-go/yatori-go-core/config/CoursesCustom.go
@@ -1,0 +1,9 @@
+package config
+
+type CoursesCustom struct {
+	VideoModel      int               `json:"videoModel"`
+	AutoExam        int               `json:"autoExam"`
+	ExcludeCourses  []string          `json:"excludeCourses"`
+	IncludeCourses  []string          `json:"includeCourses"`
+	CoursesSettings []CoursesSettings `json:"coursesSettings"`
+}
