+++ yatori-go/yatori-go-core/config/CoursesSettings.go
@@ -1,0 +1,7 @@
+package config
+
+type CoursesSettings struct {
+	Name         string   `json:"name"`
+	IncludeExams []string `json:"includeExams"`
+	ExcludeExams []string `json:"excludeExams"`
+}
