+++ yatori-go/yatori-go-core/config/EmailInform.go
@@ -1,0 +1,9 @@
+package config
+
+type EmailInform struct {
+	Sw       int    `json:"sw"`
+	SMTPHost string `json:"smtpHost"`
+	SMTPPort string `json:"smtpPort"`
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
