+++ yatori-go/yatori-go-core/config/enter.go
@@ -1,0 +1,12 @@
+package config
+
+type Config struct {
+	Setting Setting `json:"setting"`
+	Users   []Users `json:"users"`
+}
+
+type Setting struct {
+	BasicSetting BasicSetting `json:"basicSetting"`
+	EmailInform  EmailInform  `json:"emailInform"`
+	AiSetting    AiSetting    `json:"aiSetting"`
+}
