+++ yatori-go/yatori-go-core/global/global.go
@@ -1,0 +1,7 @@
+package global
+
+import "yatori-go-core/config"
+
+var (
+	Config *config.Config
+)
